% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jacobian.R
\name{calculate_jacobian}
\alias{calculate_jacobian}
\title{Calculate Jacobian Matrix}
\usage{
calculate_jacobian(C1, stoi, icount)
}
\arguments{
\item{C1}{Covariance matrix.}

\item{stoi}{Stoichiometric matrix for interactions.}

\item{icount}{Number of iterations.}
}
\value{
A list containing the Jacobian matrix and related statistics.
}
\description{
This function calculates a Jacobian matrix based on covariance data.
}
