% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_covariance.R
\name{generate_covariance}
\alias{generate_covariance}
\title{Generate Covariance Matrices for Each Time Point}
\usage{
generate_covariance(data, num_tp)
}
\arguments{
\item{data}{Numeric matrix of data.}

\item{num_tp}{Number of time points.}
}
\value{
A list of covariance matrices, one for each time point.
}
\description{
Calculates covariance matrices for subsets of data based on time points.
}
