% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_3d.R
\name{visualize_3d}
\alias{visualize_3d}
\title{Visualize Jacobian Difference in 3D}
\usage{
visualize_3d(matrix)
}
\arguments{
\item{matrix}{The matrix to visualize.}
}
\description{
Creates a 3D perspective plot for the given matrix.
}
