% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_heatmap.R
\name{visualize_heatmap}
\alias{visualize_heatmap}
\title{Visualize Jacobian Matrix as Heatmap}
\usage{
visualize_heatmap(matrix, title = "Heatmap")
}
\arguments{
\item{matrix}{The matrix to visualize.}

\item{title}{The title of the heatmap.}
}
\description{
Plots a heatmap of the given matrix with a custom color palette.
}
